Add instance member in proptypes errors
