+++ tools/analyzer_plugin/lib/src/diagnostic/proptypes_return_value.dart
@@ -1,10 +1,9 @@
 import 'package:analyzer/dart/ast/ast.dart';
-import 'package:analyzer/dart/ast/visitor.dart';
 import 'package:analyzer/source/source_range.dart';
 import 'package:analyzer_plugin/protocol/protocol_common.dart';
+import 'package:over_react_analyzer_plugin/src/diagnostic/visitors/proptypes_visitors.dart';
 import 'package:over_react_analyzer_plugin/src/diagnostic_contributor.dart';
 import 'package:over_react_analyzer_plugin/src/util/ast_util.dart';
-import 'package:over_react_analyzer_plugin/src/util/react_types.dart';
 
 /// A warning that appears when `throw`ing a `PropError` instead of returning it within `UiComponent2.propTypes`.
 ///
@@ -40,46 +39,5 @@ class PropTypesReturnValueDiagnostic extends DiagnosticContributor {
         );
       }
     }
-  }
-}
-
-class PropTypesVisitor extends SimpleAstVisitor<void> {
-  final PropTypesMapVisitor mapVisitor = PropTypesMapVisitor();
-
-  @override
-  void visitCompilationUnit(CompilationUnit node) {
-    node.visitChildren(this);
-  }
-
-  @override
-  void visitClassDeclaration(ClassDeclaration node) {
-    if (node.declaredElement.isComponentClass) {
-      node.visitChildren(this);
-    }
-  }
-
-  @override
-  void visitMethodDeclaration(MethodDeclaration node) {
-    if (node.name.name == 'propTypes') {
-      node.body.accept(mapVisitor);
-    }
-  }
-}
-
-class PropTypesMapVisitor extends RecursiveAstVisitor<void> {
-  final List<FunctionExpression> values = [];
-
-  @override
-  void visitMapLiteralEntry(MapLiteralEntry node) {
-    values.add(node.value);
-  }
-}
-
-class PropTypeFunctionBlockVisitor extends RecursiveAstVisitor<void> {
-  final List<ThrowExpression> throwExpressions = [];
-
-  @override
-  void visitExpressionStatement(ExpressionStatement node) {
-    throwExpressions.addAll(node.childEntities?.whereType<ThrowExpression>() ?? []);
   }
 }
