+++ tools/analyzer_plugin/lib/src/plugin.dart
@@ -57,6 +57,7 @@ import 'package:over_react_analyzer_plugin/src/diagnostic/callback_ref.dart';
 import 'package:over_react_analyzer_plugin/src/diagnostic/consumed_props_return_value.dart';
 import 'package:over_react_analyzer_plugin/src/diagnostic/forward_only_dom_props_to_dom_builders.dart';
 import 'package:over_react_analyzer_plugin/src/diagnostic/link_target_without_rel.dart';
+import 'package:over_react_analyzer_plugin/src/diagnostic/proptypes_instance_members.dart';
 import 'package:over_react_analyzer_plugin/src/diagnostic/proptypes_return_value.dart';
 import 'package:over_react_analyzer_plugin/src/diagnostic_contributor.dart';
 import 'package:over_react_analyzer_plugin/src/diagnostic/incorrect_doc_comment_location.dart';
@@ -148,6 +149,7 @@ class OverReactAnalyzerPlugin extends ServerPlugin
   @override
   List<DiagnosticContributor> getDiagnosticContributors(String path) {
     return [
+      PropTypesInstanceMembersDiagnostic(),
       PropTypesReturnValueDiagnostic(),
       DuplicatePropCascadeDiagnostic(),
       LinkTargetUsageWithoutRelDiagnostic(),
