+++ tools/analyzer_plugin/playground/web/proptypes.dart
@@ -10,9 +10,21 @@ mixin FooProps on UiProps {
 }
 
 class FooComponent extends UiComponent2<FooProps> {
+  String get someGetter => null;
+
+  void someMethod() {}
+  String _someField;
+
   @override
   get propTypes => {
     keyForProp((p) => p.prop1): (props, info) {
+      _someField = 'foo';
+      if (someGetter == 'foo') {
+        print('woo');
+      }
+      someMethod();
+      contextType;
+
       if (props.prop1 == null) {
         throw PropError.required(info.propName);
       } else if (props.prop1 > 0) {
@@ -21,6 +33,12 @@ class FooComponent extends UiComponent2<FooProps> {
       return null;
     },
     keyForProp((p) => p.prop2): (props, info) {
+      _someField = 'foo';
+      if (someGetter == 'foo') {
+        print('woo');
+      }
+      someMethod();
+      contextType;
       if (props.prop2 == null) {
         throw PropError.required(info.propName);
       } else if (props.prop2 > 0) {
